fix(tts): keep transcript directive parser acyclic
